moved shop comps into transition page
